# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/session": {
      "post": {
        "operationId": "AuthController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "summary": "Đăng nhập: đổi ID token lấy session cookie httpOnly",
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Đăng xuất: xoá session cookie",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "summary": "Đăng ký bằng email + mật khẩu",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "summary": "Đăng nhập bằng email hoặc số điện thoại + mật khẩu",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password/set": {
      "post": {
        "operationId": "AuthController_setPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Đặt mật khẩu lần đầu cho tài khoản chưa có (vd tạo bằng SĐT/OTP)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password/forgot": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Quên mật khẩu: gửi link đặt lại qua email",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Đặt lại mật khẩu từ token trong email",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/public/booking-requests": {
      "post": {
        "operationId": "PublicBookingRequestsController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingRequestReceiptDto"
                }
              }
            }
          }
        },
        "summary": "Khách gửi yêu cầu thuê một xe trên Marketplace",
        "tags": [
          "public-booking-requests"
        ]
      }
    },
    "/public/booking-requests/check-availability": {
      "post": {
        "operationId": "PublicBookingRequestsController_checkAvailability",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAvailabilityResultDto"
                }
              }
            }
          }
        },
        "summary": "Khách kiểm tra nhanh khung giờ của một xe có trống không (preview)",
        "tags": [
          "public-booking-requests"
        ]
      }
    },
    "/auth/phone/send-otp": {
      "post": {
        "operationId": "PhoneVerificationController_sendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendOtpResultDto"
                }
              }
            }
          }
        },
        "summary": "Gửi mã OTP tới số điện thoại",
        "tags": [
          "phone-verification"
        ]
      }
    },
    "/auth/phone/verify-otp": {
      "post": {
        "operationId": "PhoneVerificationController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOtpResultDto"
                }
              }
            }
          }
        },
        "summary": "Xác nhận mã OTP",
        "tags": [
          "phone-verification"
        ]
      }
    },
    "/auth/phone/login": {
      "post": {
        "operationId": "PhoneVerificationController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "summary": "Đăng nhập passwordless bằng SĐT + OTP (tự tạo tài khoản nếu chưa có)",
        "tags": [
          "phone-verification"
        ]
      }
    },
...
  },
  "info": {
    "title": "XePrime API",
    "description": "Convention: success `{ data, meta }`, error `{ error: { code, message, details } }`. Xác thực bằng httpOnly session cookie (ADR 0002) — không dùng Bearer token. Tiền trả về dạng string, thời gian ISO-8601 UTC (ADR 0007).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "xp_session"
      }
    },
    "schemas": {
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "ID token từ Firebase Auth (hoặc `mock:<uid>:<email>:<tên>` khi AUTH_MODE=mock). Chỉ gửi đúng một lần lúc đăng nhập — sau đó dùng session cookie.",
            "example": "mock:demo-owner:owner@xeprime.test:Chủ shop demo"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "CurrentTenantSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "Xem TenantStatus trong @xeprime/types"
          },
          "roleKey": {
            "type": "string",
            "description": "Xem TenantRole trong @xeprime/types"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "status",
          "roleKey"
        ]
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "phoneVerified": {
            "type": "boolean",
            "description": "Đã xác thực SĐT chưa — gate cho việc đặt xe/mở shop"
          },
          "hasPassword": {
            "type": "boolean",
            "description": "Đã có mật khẩu chưa — false với tài khoản tạo bằng SĐT/OTP (gợi ý đặt mật khẩu)"
          },
          "tenant": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CurrentTenantSummaryDto"
              }
            ]
          },
          "platformRole": {
            "type": "object",
            "nullable": true,
            "description": "Xem PlatformRole trong @xeprime/types"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "phoneVerified",
          "hasPassword",
          "permissions"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "matkhau123"
          },
          "displayName": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "email": {
            "type": "string",
            "example": "ban@congty.vn"
          }
        },
        "required": [
          "password",
          "displayName",
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Email hoặc số điện thoại",
            "example": "ban@congty.vn"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "matkhau123"
          }
        },
        "required": [
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ban@congty.vn"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "matkhau123"
          },
          "token": {
            "type": "string",
            "description": "Token từ link trong email"
          }
        },
        "required": [
          "password",
          "token"
        ]
      },
...
      "CreateBookingRequestDto": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "description": "ID xe (ULID) trên marketplace"
          },
          "customerName": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "customerPhone": {
            "type": "string",
            "example": "0901234567"
          },
          "customerEmail": {
            "type": "string",
            "example": "a@example.com"
          },
          "pickupAt": {
            "type": "string",
            "description": "Nhận xe (ISO-8601)"
          },
          "returnAt": {
            "type": "string",
            "description": "Trả xe (ISO-8601), phải sau nhận xe"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "vehicleId",
          "customerName",
          "customerPhone",
          "pickupAt",
          "returnAt"
        ]
      },
      "BookingRequestReceiptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_host_approval",
              "approved_by_host",
              "rejected_by_host",
              "cancelled_by_customer",
              "expired",
              "converted_to_booking"
            ]
          },
          "authenticated": {
            "type": "boolean",
            "description": "Sau khi gửi, khách đã có phiên đăng nhập (passwordless qua SĐT) — FE chuyển tới /trips."
          }
        },
        "required": [
          "id",
          "status",
          "authenticated"
        ]
      },
      "CheckAvailabilityDto": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "string",
            "description": "ID xe (ULID)"
          },
          "pickupAt": {
            "type": "string",
            "description": "Nhận xe (ISO-8601)"
          },
          "returnAt": {
            "type": "string",
            "description": "Trả xe (ISO-8601), phải sau nhận xe"
          }
        },
        "required": [
          "vehicleId",
          "pickupAt",
          "returnAt"
        ]
      },
      "CheckAvailabilityResultDto": {
        "type": "object",
        "properties": {
          "available": {
            "type": "boolean",
            "description": "Khung giờ còn trống (preview, quyết định thật khi shop duyệt)"
          }
        },
        "required": [
          "available"
        ]
      },
      "SendOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "0901234567"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "booking",
              "login",
              "shop_register",
              "vehicle_public",
              "password_reset"
            ],
            "example": "booking"
          }
        },
        "required": [
          "phone",
          "purpose"
        ]
      },
      "SendOtpResultDto": {
        "type": "object",
        "properties": {
          "expiresAt": {
            "type": "string",
            "description": "ISO-8601 UTC — thời điểm mã hết hạn"
          },
          "devCode": {
            "type": "string",
            "nullable": true,
            "description": "CHỈ ở dev (OTP_MODE=mock): mã để tự điền/test. Production luôn null."
          }
        },
        "required": [
          "expiresAt"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "0901234567"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "booking",
              "login",
              "shop_register",
              "vehicle_public",
              "password_reset"
            ],
            "example": "booking"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "phone",
          "purpose",
          "code"
        ]
      },
      "VerifyOtpResultDto": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean"
          }
        },
        "required": [
          "verified"
        ]
      },
      "PhoneLoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "0901234567"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
...
    }
  }
}